+++ CHANGELOG.md
@@ -8,9 +8,11 @@ This changelog only contains the changes that are unreleased. For changes for in
 ### New Features
 - Allow skipping browser download mods for cases where mods no longer exist
 - Show progress of downloading browser mods
+- Add new dialog when downloading browser mods is needed for modpacks to make it easier
 
 ### Fixes
 - Custom downloads folders in Windows not being used for browser downloads [#586]
 - When using Flatpak, only show the launchers downloads folder and not the users one
+- Instances not becoming corrupt anymore when installs fail
 
 ### Misc